+++ adapters/Facebook/FacebookAdapter/GADFBAdapterDelegate.m
@@ -194,7 +194,7 @@
   [strongConnector adapter:strongAdapter didFailToLoadRewardBasedVideoAdwithError:error];
 }
 
-- (void)rewardedVideoAdComplete:(FBRewardedVideoAd *)rewardedVideoAd {
+- (void)rewardedVideoAdVideoComplete:(FBRewardedVideoAd *)rewardedVideoAd {
   id<GADMRewardBasedVideoAdNetworkConnector> strongConnector = _rewardBasedVideoAdConnector;
   id<GADMRewardBasedVideoAdNetworkAdapter> strongAdapter = _rewardBasedVideoAdAdapter;
   [strongConnector adapter:strongAdapter