+++ adapters/IronSource/Public/Headers/IronSourceAdapter.h
@@ -12,8 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#ifndef IronSourceAdapter_h
+#define IronSourceAdapter_h
+
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
+#import <IronSourceAdapter/GADMIronSourceExtras.h>
 
-/// Import all your public header files here.
-//#import <IronSourceAdapter/YOUR_PUBLIC_HEADER.h>
+#endif /* IronSourceAdapter_h */