+++ adapters/AppLovin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # AppLovin Adapter for Google Mobile Ads SDK for iOS
 
+## Version 5.0.1.0
+- Verified compatibility with Applovin SDK 5.0.1.
+
+## Version 4.8.4.0
+- Verified compatibility with Applovin SDK 4.8.4.
+
+## Version 4.8.3.0
+- Add support for zones and smart banners.
+
+## Version 4.7.0.0
+- Verified compatibility with AppLovin SDK 4.7.0.
+
 ## Version 4.6.1.0
 - Verified compatibility with AppLovin SDK 4.6.1.